mangacross: move & create tests
