+++ web/src/engine/websites/_index.ts
@@ -173,6 +173,7 @@ export { default as MangaChan } from './MangaChan';
 export { default as Mangacim } from './Mangacim';
 export { default as MangaClash } from './MangaClash';
 export { default as MangaCrab } from './MangaCrab';
+export { default as MangaCross } from './MangaCross';
 export { default as MangaDenizi } from './MangaDenizi';
 export { default as MangaDex } from './MangaDex';
 export { default as MangaDiyari } from './MangaDiyari';
@@ -594,7 +595,6 @@ export { default as Manga99 } from './legacy/Manga99';
 export { default as Manga1001 } from './legacy/Manga1001';
 export { default as MangaArab } from './legacy/MangaArab';
 export { default as MangaCat } from './legacy/MangaCat';
-export { default as MangaCross } from './legacy/MangaCross';
 export { default as MangaCruzers } from './legacy/MangaCruzers';
 export { default as MangaDeep } from './legacy/MangaDeep';
 export { default as MangaEdenEN } from './legacy/MangaEdenEN';
